requirements.txt
discord.py==1.2.5
numpy==1.15.3